fix: various clang tidy fixes

+++ src/lib/obfuscator/transforms/transforms/decomp_break.hpp
@@ -14,7 +14,7 @@ namespace obfuscator::transforms {
     template <pe::any_image_t Img>
     class DecompBreak final : public BBTransform<Img> {
     public:
-        enum Var : std::size_t {
+        enum Var : std::uint8_t {
             BREAK_IDA = 0,
             BREAK_GHIDRA = 1,
         };
@@ -28,7 +28,7 @@ namespace obfuscator::transforms {
         /// \brief Transform zasm node
         /// \param function Routine that it should transform
         /// \param bb BB that it should transform
-        void run_on_bb(TransformContext&, Function<Img>* function, analysis::bb_t* bb) override {
+        void run_on_bb(TransformContext& /*ctx*/, Function<Img>* function, analysis::bb_t* bb) override {
             /// No successors?
             if (bb->successors.empty()) {
                 return;
